+++ Assets/Mirror/Editor/Weaver/ArraySegmentWriter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c15aa0cda824624f86a65edb3be2ace
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
