+++ Assets/Mirror/Editor/Weaver/ArrayWriter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2e3c3d6e421e7a42a6c6490ccb30775
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
