+++ Assets/Mirror/Editor/Weaver/CollectionWriter.cs
@@ -3,116 +3,6 @@ using Mono.CecilX.Cil;
 
 namespace Mirror.Weaver
 {
-    public class ArrayWriter : CollectionWriter
-    {
-        /// <exception cref="GenerateWriterException">Throws when writer could not be generated for type</exception>
-        public ArrayWriter(TypeReference variable, int recursionCount) : base(variable, recursionCount) { }
-
-        protected override string namePrefix => "Array";
-        protected override bool needNullCheck => true;
-
-        protected override TypeReference GetElementType(TypeReference variable)
-        {
-            return variable.GetElementType();
-        }
-
-        /// <exception cref="GenerateWriterException">Throws when writer could not be generated for type</exception>
-        protected override void ValidateType(TypeReference variable)
-        {
-            if (!variable.IsArrayType())
-            {
-                throw new GenerateWriterException($"{variable.Name} is an unsupported type. Jagged and multidimensional arrays are not supported", variable);
-            }
-        }
-
-        protected override Instruction CreateLengthInstruction(ILProcessor worker)
-        {
-            return worker.Create(OpCodes.Ldlen);
-        }
-    }
-    public class ArraySegmentWriter : CollectionWriter
-    {
-        GenericInstanceType genericInstance;
-
-        /// <exception cref="GenerateWriterException">Throws when writer could not be generated for type</exception>
-        public ArraySegmentWriter(TypeReference variable, int recursionCount) : base(variable, recursionCount) { }
-
-        protected override string namePrefix => "ArraySegment";
-        protected override bool needNullCheck => false;
-
-        protected override TypeReference GetElementType(TypeReference variable)
-        {
-            genericInstance = (GenericInstanceType)variable;
-            TypeReference elementType = genericInstance.GenericArguments[0];
-            return elementType;
-        }
-
-        protected override Instruction CreateLengthInstruction(ILProcessor worker)
-        {
-            MethodReference countref = WeaverTypes.ArraySegmentCountReference.MakeHostInstanceGeneric(genericInstance);
-
-            return worker.Create(OpCodes.Call, countref);
-        }
-
-        protected override void AppendCollection(ILProcessor worker)
-        {
-            MethodReference getArray = WeaverTypes.ArraySegmentArrayReference.MakeHostInstanceGeneric(genericInstance);
-
-            // arg1 is the segment
-            worker.Append(worker.Create(OpCodes.Ldarg_1));
-            // value.Array
-            worker.Append(worker.Create(OpCodes.Call, getArray));
-        }
-
-        protected override void AppendIndex(ILProcessor worker)
-        {
-            MethodReference getOffset = WeaverTypes.ArraySegmentOffsetReference.MakeHostInstanceGeneric(genericInstance);
-
-            // local 1 is i
-            worker.Append(worker.Create(OpCodes.Ldloc_1));
-
-            // value.Offset
-            worker.Append(worker.Create(OpCodes.Ldarg_1));
-            worker.Append(worker.Create(OpCodes.Call, getOffset));
-
-            // add (i + offset)
-            worker.Append(worker.Create(OpCodes.Add));
-        }
-    }
-    public class ListWriter : CollectionWriter
-    {
-        GenericInstanceType genericInstance;
-
-        /// <exception cref="GenerateWriterException">Throws when writer could not be generated for type</exception>
-        public ListWriter(TypeReference variable, int recursionCount) : base(variable, recursionCount) { }
-
-        protected override string namePrefix => "List";
-        protected override bool needNullCheck => true;
-
-        protected override TypeReference GetElementType(TypeReference variable)
-        {
-            genericInstance = (GenericInstanceType)variable;
-            TypeReference elementType = genericInstance.GenericArguments[0];
-            return elementType;
-        }
-
-        protected override Instruction CreateLengthInstruction(ILProcessor worker)
-        {
-            MethodReference countref = WeaverTypes.ListCountReference.MakeHostInstanceGeneric(genericInstance);
-
-            return worker.Create(OpCodes.Call, countref);
-        }
-
-        protected override void AppendGetElement(ILProcessor worker)
-        {
-            MethodReference getItem = WeaverTypes.ListGetItemReference.MakeHostInstanceGeneric(genericInstance);
-
-            // call get_Item
-            // list and index should be already on stack
-            worker.Append(worker.Create(OpCodes.Call, getItem));
-        }
-
-    }
     public abstract class CollectionWriter
     {
         private readonly TypeReference variable;
