+++ Assets/Mirror/Editor/Weaver/ListWriter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12173025b8a22d94e8e23ab8dfbb1910
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
